Develop

+++ app/Http/Controllers/UnitController.php
@@ -24,6 +24,18 @@ class UnitController extends Controller
     }
 
     /**
+     * Exibe o mapa com as unidades.
+     */
+    public function map()
+    {
+        $units = Unit::all();
+
+        return Inertia::render('Units/Map', [
+            'units' => $units,
+        ]);
+    }
+
+    /**
      * Show the form for creating a new resource.
      */
     public function create()
@@ -168,6 +180,9 @@ class UnitController extends Controller
         return redirect()->back()->with('success', 'Unidade atualizada com sucesso.');
     }
 
+    /**
+     * Atualiza a hierarquia das unidades.
+     */
     public function updateHierarchy(Request $request)
     {
         foreach ($request->units as $update) {
@@ -188,6 +203,18 @@ class UnitController extends Controller
     public function destroy(string $id)
     {
         $unit = Unit::findOrFail($id);
+
+        // Verifica se a unidade é uma unidade raiz
+        if ($unit->parent_id === null) {
+            return redirect()->back()->with('error', 'A unidade raiz não pode ser removida.');
+        }
+
+        // Verifica se a unidade possui filhos
+        $childUnits = $unit->children;
+        if ($childUnits->isNotEmpty()) {
+            // Se tiver, altera o parent_id delas para um novo valor
+            $unit->children()->update(['parent_id' => $unit->parent_id]);
+        }
         $unit->delete();
 
         return redirect()->back()->with('success', 'Unidade removida com sucesso.');
